+++ src/FileCachePlugin.php
@@ -49,7 +49,6 @@ class FileCachePlugin extends Plugin
 		]);
 
 		$this->initEvents();
-		$this->injectJsCsrfToken();
 	}
 
 	public function fileCacheService(): FileCacheService
@@ -151,31 +150,5 @@ class FileCachePlugin extends Plugin
 
 			$this->_deleteCaches = false;
 		}
-	}
-
-	protected function injectJsCsrfToken(): void
-	{
-		/** @var SettingsModel $settings */
-		$settings = $this->getSettings();
-
-		if (!$settings->injectJsCsrfToken) {
-			return;
-		}
-
-		$url = '/' . Craft::$app->getConfig()->getGeneral()->actionTrigger . '/filecache/csrf/js';
-		$view = Craft::$app->getView();
-		$view->registerJs(<<<Js
-var xhr = new XMLHttpRequest();
-xhr.onload = function () {
-    if (xhr.status >= 200 && xhr.status < 300) {
-        window.csrfTokenName = this.response.csrfTokenName;
-	    window.csrfTokenValue = this.response.csrfTokenValue;
-    }
-};
-xhr.open('GET', '$url');
-xhr.responseType = 'json';
-xhr.send();
-Js
-			, View::POS_END);
 	}
 }